обнова

+++ 1125/Model/Product.cs
@@ -12,6 +12,6 @@ namespace _1125.Model
         public string Name { get; set; }
         public string Description { get; set; }
         public int Availability {  get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
     }
 }
